% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh-package.R
\docType{package}
\name{gh-package}
\alias{gh-package}
\title{gh: 'GitHub' 'API'}
\description{
Minimal client to access the 'GitHub' 'API'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gh.r-lib.org/}
  \item \url{https://github.com/r-lib/gh}
  \item Report bugs at \url{https://github.com/r-lib/gh/issues}
}

}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com} [contributor]

Authors:
\itemize{
  \item Jennifer Bryan
  \item Hadley Wickham
}

Other contributors:
\itemize{
  \item Posit Software, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
}

}
\keyword{internal}
