% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh-apps.R
\name{gh_app_token}
\alias{gh_app_token}
\title{Authenticate as an App}
\usage{
gh_app_token(
  installation,
  app_id = Sys.getenv("GH_APP_ID"),
  app_key = Sys.getenv("GH_APP_KEY")
)
}
\arguments{
\item{installation}{the target app-installation to authenticate with, This must
either be a user / organization name such as \code{"ropensci"}, or a repository that
has the app installed for example \code{"ropensci/magick"}.}

\item{app_id}{a string with the github app id}

\item{app_key}{file or string with your private key, passed to \link[openssl:read_key]{openssl::read_key}}
}
\value{
a temporary token that will be valid for 1 hour
}
\description{
Authenticates with the GitHub API on behalf of a GitHub app using the
app-id and private-key file that you get when registering the app. The
result is a temporary token that is valid for 1 hour, which you can use
just like other tokens to act on behalf of the app using e.g. \code{\link[=gh]{gh()}}.
}
\details{
To register a new app go to:
\url{https://github.com/settings/apps/new}
}
