% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_gql.R
\name{gh_gql}
\alias{gh_gql}
\title{A simple interface for the GitHub GraphQL API v4.}
\usage{
gh_gql(
  query,
  ...,
  .token = NULL,
  .destfile = NULL,
  .overwrite = FALSE,
  .api_url = NULL,
  .limit = NULL,
  .send_headers = NULL
)
}
\arguments{
\item{query}{The GraphQL query, as a string.}

\item{...}{Name-value pairs giving API parameters. Will be matched
into \code{endpoint} placeholders, sent as query parameters in GET
requests, and as a JSON body of POST requests. Named \code{NULL} values
are silently dropped, and named \code{NA} values trigger an error.}

\item{.token}{Authentication token. Defaults to \code{GITHUB_PAT} or
\code{GITHUB_TOKEN} environment variables, in this order if any is set.}

\item{.destfile}{path to write response to disk.  If NULL (default), response will
be processed and returned as an object.  If path is given, response will
be written to disk in the form sent.}

\item{.overwrite}{if \code{.destfile} is provided, whether to overwrite an
existing file.  Defaults to FALSE.}

\item{.api_url}{Github API url (default: \url{https://api.github.com}). Used
if \code{endpoint} just contains a path. Defaults to \code{GITHUB_API_URL}
environment variable if set.}

\item{.limit}{Number of records to return. This can be used
instead of manual pagination. By default it is \code{NULL},
which means that the defaults of the GitHub API are used.
You can set it to a number to request more (or less)
records, and also to \code{Inf} to request all records.
Note, that if you request many records, then multiple GitHub
API calls are used to get them, and this can take a potentially
long time.}

\item{.send_headers}{Named character vector of header field values
(except \code{Authorization}, which is handled via \code{.token}). This can be
used to override or augment the default \code{User-Agent} header:
\code{"https://github.com/r-lib/gh"}.}
}
\description{
A simple interface for the GitHub GraphQL API v4.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
gh_gql("query { viewer { login }}")
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=gh]{gh()}} for the GitHub v3 API.
}
