% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_request.R
\name{gh_interactive_oauth_token}
\alias{gh_interactive_oauth_token}
\title{Obtain a Github OAuth Token Interactively}
\usage{
gh_interactive_oauth_token(client_id, client_secret, github_url)
}
\arguments{
\item{client_id}{The client ID you received from GitHub when you
\href{https://github.com/settings/applications/new}{registered} your Github app.}

\item{client_secret}{The client secret you received from GitHub for your GitHub App.}

\item{github_url}{The base url of Github or Github Enterprise.}
}
\value{
An \code{\link[httr:oauth2.0_token]{httr::oauth2.0_token()}}
}
\description{
Under an interactive R session, this opens the browser and acquires an OAuth
token interactively. This provides a simplified authorization experience than
manually creating PAT and setting the environment variable.
}
\details{
If \code{client_id} and \code{client_secret} are not provided, a default pair is used
for github.com. Refer to \href{https://developer.github.com/apps/building-oauth-apps/creating-an-oauth-app/}{Github documentation}
on how to create your own application. The callback URL shall be the value of
\code{\link[httr:oauth_callback]{httr::oauth_callback()}}, namely, \code{http://localhost:1410/}.

This can be useful for Github Enterprise installation where an internal
R package can create a Github application such that users may obtain authroization
tokens interactively. The package may further wrap this token in helper functions
such as streamlining internal \code{install_github} usage. See examples section for more details.
}
\examples{
\dontrun{
# obtain an OAuth token with GHE
token <- withr::with_options(
  # share GHE OAuth token across projects
  httr_oauth_cache = "~/.httr_oauth_ghe",
  gh_interactive_oauth_token("{client_id}", "{client_secret}", "{ghe_host}")
)
# use obtained OAuth token temporarily
withr::with_envvar(
  GITHUB_PAT = token$credentials$access_token,
  gh_whoami()
)
# wrap it to create custom install_github function
install_github <- function(...) {
  devtools::install_github(host = "{ghe_api}", auth_token = token$credentials$access_token, ...)
}
}
}
\references{
\link{http://developer.github.com/guides/basics-of-authentication/}
}
