% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{gh_tree_remote}
\alias{gh_tree_remote}
\title{Find the GitHub remote associated with a path}
\usage{
gh_tree_remote(path = ".")
}
\arguments{
\item{path}{Path that is contained within a git repo.}
}
\value{
If the repo has a github remote, a list containing \code{username}
and \code{repo}. Otherwise, an error.
}
\description{
This is handy helper if you want to make gh requests related to the
current project.
}
\examples{
\dontrun{
gh_tree_remote()
}
}
