% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.gh_response}
\alias{print.gh_response}
\title{Print the result of a GitHub API call}
\usage{
\method{print}{gh_response}(x, ...)
}
\arguments{
\item{x}{The result object.}

\item{...}{Ignored.}
}
\value{
The JSON result.
}
\description{
Print the result of a GitHub API call
}
