% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_token.R
\name{slugify_url}
\alias{slugify_url}
\title{Compute the suffix that gh uses for GitHub API URL specific PATs}
\usage{
slugify_url(url)
}
\arguments{
\item{url}{Character vector of HTTP/HTTPS URLs. They don't have to be in the
API-specific form, although they can be.}
}
\value{
Character vector of suffixes.
}
\description{
\code{slugify_url()} determines a suffix from a URL and this suffix is used to
construct the name of an environment variable that holds the PAT for a
specific GitHub URL. This is mostly relevant to people using GitHub
Enterprise.
}
\examples{
# main github.com site
slugify_url("https://api.github.com")
slugify_url("https://github.com")

# an instance of GitHub Enterprise
slugify_url("https://github.acme.com")
slugify_url("https://github.acme.com/api/v3")
}
\seealso{
\code{\link[=gh_token]{gh_token()}}
}
