% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{gh}
\alias{gh}
\alias{gh-package}
\title{GitHub API}
\usage{
gh(end_point, ..., .token = Sys.getenv("GITHUB_TOKEN"))
}
\arguments{
\item{end_point}{GitHub API end point. See examples below.}

\item{...}{Additional parameters}

\item{.token}{Authentication token.}
}
\value{
Answer from the API.
}
\description{
Minimal wrapper to access GitHub's API.

TODO
}
\examples{
\dontrun{
## Repositories of a user, these are equivalent
gh("/users/hadley/repos")
gh("/users/:username/repos", username = "hadley")

## Create a repository
gh("POST /user/repos", name = "foobar")

## Issues of a repository
gh("/repos/hadley/dplyr/issues")
gh("/repos/:owner/:repo/issues", owner = "hadley", repo = "dplyr")
}
}

