% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{gh}
\alias{gh}
\alias{gh-package}
\title{GitHub API}
\usage{
gh(end_point, ...)
}
\arguments{
\item{end_point}{GitHub API end point. See examples below.}

\item{...}{Additional parameters}

\item{.token}{Authentication token.}
}
\value{
Answer from the API.
}
\description{
Minimal wrapper to access GitHub's API.

TODO
}
\examples{
\dontrun{
## Repositories of a user
gh(users/hadley/repos)

## Issues of a repository
gh(repos/hadley/dplyr/issues)

## Using variables in the end point, just put them in a paren
myuser <- "gaborcsardi"
gh(users/(myuser)/repos)
}
}

