% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pagination.R
\name{gh_next}
\alias{gh_first}
\alias{gh_last}
\alias{gh_next}
\alias{gh_prev}
\title{Get the next, previous, first or last page of results}
\usage{
gh_next(gh_response, .token = Sys.getenv("GITHUB_TOKEN"),
  .send_headers = NULL)

gh_prev(gh_response, .token = Sys.getenv("GITHUB_TOKEN"),
  .send_headers = NULL)

gh_first(gh_response, .token = Sys.getenv("GITHUB_TOKEN"),
  .send_headers = NULL)

gh_last(gh_response, .token = Sys.getenv("GITHUB_TOKEN"),
  .send_headers = NULL)
}
\arguments{
\item{gh_response}{An object returned by a \code{gh()} call.}

\item{.token}{Authentication token.}

\item{.send_headers}{Named character vector of header field values
  (excepting \code{Authorization}, which is handled via
  \code{.token}). This can be used to override or augment the
  defaults, which are as follows: the \code{Accept} field defaults
  to \code{"application/vnd.github.v3+json"} and the
  \code{User-Agent} field defaults to
  \code{"https://github.com/gaborcsardi/whoami"}. This can be used
  to, e.g., provide a custom media type, in order to access a
  preview feature of the API.}
}
\value{
Answer from the API.
}
\description{
Get the next, previous, first or last page of results
}
\details{
Note that these are not always defined. E.g. if the first
page was queried (the default), then there are no first and previous
pages defined. If there is no next page, then there is no
next page defined, etc.

If the requested page does not exist, an error is thrown.
}
\examples{
\dontrun{
x <- gh("/users")
sapply(x, "[[", "login")
x2 <- gh_next(x)
sapply(x2, "[[", "login")
}
}

