% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_token.R
\name{slugify_url}
\alias{slugify_url}
\title{Compute the suffix that gh uses for GitHub API URL specific PATs}
\usage{
slugify_url(url)
}
\arguments{
\item{url}{Character vector HTTP/HTTPS URLs.}
}
\value{
Character vector of suffixes.
}
\description{
Compute the suffix that gh uses for GitHub API URL specific PATs
}
\examples{
# The main GH site
slugify_url("https://api.github.com")

# A custom one
slugify_url("https://github.acme.com")
}
\seealso{
\code{\link[=gh_token]{gh_token()}}
}
